Support initial graphics API switching
