+++ projects/msvc/GLideN64.vcxproj.filters
@@ -775,6 +775,9 @@
     <ClInclude Include="..\..\src\QueueExecutor.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="..\..\src\ConfigDecl.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="..\..\src\GLideNUI-wtl\GLideNUI.rc" />
