+++ src/GLideNUI-wtl/UIConfig.cpp
@@ -1,0 +1,9 @@
+#include "UIConfig.h"
+
+#include "../PluginAPI.h"
+#include "../wst.h"
+
+namespace UI
+{
+	Config config;
+}
