+++ src/GLideNUI-wtl/UIConfig.h
@@ -1,0 +1,11 @@
+#ifndef CONFIG_H
+#define CONFIG_H
+
+#include "../ConfigDecl.h"
+
+inline namespace UI
+{
+extern Config config;
+}
+
+#endif // CONFIG_H
