+++ src/GLideNUI/GLideNUI.h
@@ -1,25 +1,10 @@
 #ifndef GLIDENUII_H
 #define GLIDENUII_H
 
-#if defined(__cplusplus)
-extern "C" {
-#endif
-
-#ifdef OS_WINDOWS
-#define EXPORT	__declspec(dllexport)
-#define CALL		__cdecl
-#else
-#define EXPORT 	__attribute__((visibility("default")))
-#define CALL
-#endif
-
-EXPORT bool CALL RunConfig(const wchar_t * _strFileName, const char * _romName, unsigned int _maxMSAALevel, unsigned int _maxAnisotropy);
-EXPORT int CALL RunAbout(const wchar_t * _strFileName);
-EXPORT void CALL LoadConfig(const wchar_t * _strFileName);
-EXPORT void CALL LoadCustomRomSettings(const wchar_t * _strFileName, const char * _romName);
-
-#if defined(__cplusplus)
-}
-#endif
+class Config;
+bool RunConfig(const wchar_t * _strFileName, const char * _romName, unsigned int _maxMSAALevel, unsigned int _maxAnisotropy);
+int RunAbout(const wchar_t * _strFileName);
+void LoadConfig(Config*, const wchar_t * _strFileName);
+void LoadCustomRomSettings(Config*, const wchar_t * _strFileName, const char * _romName);
 
 #endif // GLIDENUII_H
